Add db opening
