+++ .env
@@ -1,2 +1,4 @@
 port: :5000
-BAR: baz
+dbUser: 
+dbPassword: 
+dbName:
